MCP_Test_Client/pyproject.toml
[project]
name = "mcp-test-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.55.0",
    "mcp>=1.10.1",
    "python-dotenv>=1.1.1",
]
